Avoid unportable left shift construct
left shift of 9 by 28 places cannot be represented in type 'int'

+++ lib/libc/gdtoa/gethex.c
@@ -1,4 +1,4 @@
-/* $NetBSD: gethex.c,v 1.6 2013/04/19 10:41:53 joerg Exp $ */
+/* $NetBSD: gethex.c,v 1.7 2020/02/22 00:38:14 kamil Exp $ */
 
 /****************************************************************
 
@@ -209,7 +209,7 @@ gethex( CONST char **sp, CONST FPI *fpi, Long *expt, Bigint **bp, int sign, loca
 			L = 0;
 			n = 0;
 			}
-		L |= (hexdig[(unsigned char)*s1] & 0x0f) << n;
+		L |= (unsigned int)(hexdig[(unsigned char)*s1] & 0x0f) << n;
 		n += 4;
 		}
 	*x++ = L;
